nuevas actualizaciones

+++ src/main/java/com/codigo/spring/repository/VueloRepository.java
@@ -3,5 +3,5 @@ package com.codigo.spring.repository;
 import com.codigo.spring.entity.VueloEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VueloRepository extends JpaRepository<VueloEntity, Long> {
+public interface VueloRepository extends JpaRepository<VueloEntity, Integer> {
 }
